End practices use reducer!
