+++ src/reducers/hooks/index.js
@@ -1,1 +1,2 @@
 export * from './useForm';
+export * from './useTarea';
